:recycle: Judgement 관련 패키지 분리
