--- src/main/java/kr/co/morandi/backend/defense_management/infrastructure/baekjoon/judgement/submit/BaekjoonJudgementConstants.java
@@ -1,9 +1,0 @@
-package kr.co.morandi.backend.defense_management.infrastructure.baekjoon.judgement.submit;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class BaekjoonJudgementConstants {
-    public static final String BAEKJOON_USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36";
-    public static final String BAEKJOON_SUBMIT_URL = "https://www.acmicpc.net/submit/%s";
-}
