+++ src/main/java/kr/co/morandi/backend/judgement/domain/error/BaekjoonCookieErrorCode.java
@@ -1,0 +1,25 @@
+package kr.co.morandi.backend.judgement.domain.error;
+
+import kr.co.morandi.backend.common.exception.errorcode.ErrorCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@RequiredArgsConstructor
+public enum BaekjoonCookieErrorCode implements ErrorCode {
+
+    INVALID_COOKIE_VALUE(HttpStatus.BAD_REQUEST, "쿠키 값은 null이거나 빈 문자열일 수 없습니다.");
+
+    private final HttpStatus httpStatus;
+    private final String message;
+
+    @Override
+    public HttpStatus getHttpStatus() {
+        return httpStatus;
+    }
+    @Override
+    public String getMessage() {
+        return message;
+    }
+}
