+++ src/main/java/kr/co/morandi/backend/judgement/infrastructure/baekjoon/submit/BaekjoonSubmitConstants.java
@@ -1,0 +1,9 @@
+package kr.co.morandi.backend.judgement.infrastructure.baekjoon.submit;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class BaekjoonSubmitConstants {
+    public static final String BAEKJOON_USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36";
+    public static final String BAEKJOON_SUBMIT_URL = "https://www.acmicpc.net/submit/%s";
+}
