+++ src/test/java/kr/co/morandi/backend/defense_management/infrastructure/adapter/judgement/BaekjoonJudgementConstantsTest.java
@@ -1,6 +1,6 @@
 package kr.co.morandi.backend.defense_management.infrastructure.adapter.judgement;
 
-import kr.co.morandi.backend.defense_management.infrastructure.baekjoon.judgement.submit.BaekjoonJudgementLanguageCode;
+import kr.co.morandi.backend.judgement.infrastructure.baekjoon.submit.BaekjoonSubmitLanguageCode;
 import kr.co.morandi.backend.defense_management.domain.model.tempcode.model.Language;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,7 +16,7 @@ class BaekjoonJudgementConstantsTest {
         Language language = Language.CPP;
 
         // when
-        String languageId = BaekjoonJudgementLanguageCode.getLanguageCode(language);
+        String languageId = BaekjoonSubmitLanguageCode.getLanguageCode(language);
 
         // then
         assertThat(languageId)
